% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseVFC.R
\name{predict.VFC}
\alias{predict.VFC}
\title{Predict method for VFC fits}
\usage{
\method{predict}{VFC}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{VFC} object generated by \code{\link[=SparseVFC]{SparseVFC()}}.}

\item{newdata}{A vector specifying the position.}

\item{...}{Not in use.}
}
\value{
A vector.
}
\description{
Predicted values based on \code{VFC} objects.
}
\seealso{
\code{\link[=SparseVFC]{SparseVFC()}}
}
